requirements.txt
Flask
google-genai
pydantic
python-dotenv
pymongo
flask-corsx